Create geometry package
