+++ library/src/main/java/com/otaliastudios/opengl/geometry/IndexedPointF.kt
@@ -1,0 +1,5 @@
+package com.otaliastudios.opengl.geometry
+
+import android.graphics.PointF
+
+class IndexedPointF(val index: Int, x: Float, y: Float) : PointF(x, y)
